test_deepFace/requirements.txt
deepface
opencv-python
matplotlib
tf-keras
tensorflow-addons
numpy
pandas
scipy
Pillow
tqdm
gdown